+++ examples/ai_guard/ai_guard_examples/guard.py
@@ -15,4 +15,11 @@ input_text = "This email address, security@pangea.cloud, will be redacted."
 print("Guarding text:", input_text)
 text_response = ai_guard.guard_text(input_text)
 assert text_response.result
-print("Response:", text_response.result.prompt)
+print("Response:", text_response.result.prompt_text)
+
+# Structured input.
+structured_input = [{"role": "user", "content": "hello world"}]
+print("Guarding structured input:", structured_input)
+structured_response = ai_guard.guard_text(structured_input)
+assert structured_response.result
+print("Response:", structured_response.result.prompt_messages)