+++ packages/pangea-sdk/pangea/asyncio/services/ai_guard.py
@@ -1,9 +1,15 @@
 from __future__ import annotations
 
+from typing import overload
+
+from typing_extensions import TypeVar
+
 from pangea.asyncio.services.base import ServiceBaseAsync
 from pangea.config import PangeaConfig
 from pangea.response import PangeaResponse
 from pangea.services.ai_guard import TextGuardResult
+
+_T = TypeVar("_T")
 
 
 class AIGuardAsync(ServiceBaseAsync):
@@ -45,31 +51,106 @@ class AIGuardAsync(ServiceBaseAsync):
 
         super().__init__(token, config, logger_name, config_id)
 
+    @overload
     async def guard_text(
         self,
-        text: str,
+        text_or_messages: str,
         *,
         recipe: str = "pangea_prompt_guard",
         debug: bool = False,
-    ) -> PangeaResponse[TextGuardResult]:
+    ) -> PangeaResponse[TextGuardResult[None]]:
         """
-        Text guard (Beta)
+        Text Guard for scanning LLM inputs and outputs (Beta)
 
-        Guard text.
+        Analyze and redact text to avoid manipulation of the model, addition of
+        malicious content, and other undesirable data transfers.
 
         How to install a [Beta release](https://pangea.cloud/docs/sdk/python/#beta-releases).
 
         OperationId: ai_guard_post_v1beta_text_guard
 
         Args:
-            text: Text.
-            recipe: Recipe.
-            debug: Debug.
+            text: Text to be scanned by AI Guard for PII, sensitive data,
+                malicious content, and other data types defined by the
+                configuration. Supports processing up to 10KB of text.
+            recipe: Recipe key of a configuration of data types and settings
+                defined in the Pangea User Console. It specifies the rules that
+                are to be applied to the text, such as defang malicious URLs.
+            debug: Setting this value to true will provide a detailed analysis
+                of the text data
 
         Examples:
             response = await ai_guard.guard_text("text")
         """
 
+    @overload
+    async def guard_text(
+        self,
+        text_or_messages: _T,
+        *,
+        recipe: str = "pangea_prompt_guard",
+        debug: bool = False,
+    ) -> PangeaResponse[TextGuardResult[_T]]:
+        """
+        Text Guard for scanning LLM inputs and outputs (Beta)
+
+        Analyze and redact text to avoid manipulation of the model, addition of
+        malicious content, and other undesirable data transfers.
+
+        How to install a [Beta release](https://pangea.cloud/docs/sdk/python/#beta-releases).
+
+        OperationId: ai_guard_post_v1beta_text_guard
+
+        Args:
+            text_or_messages: Structured data to be scanned by AI Guard for PII,
+                sensitive data, malicious content, and other data types defined
+                by the configuration. Supports processing up to 10KB of text.
+            recipe: Recipe key of a configuration of data types and settings
+                defined in the Pangea User Console. It specifies the rules that
+                are to be applied to the text, such as defang malicious URLs.
+            debug: Setting this value to true will provide a detailed analysis
+                of the text data
+
+        Examples:
+            response = await ai_guard.guard_text([
+                {"role": "user", "content": "hello world"}
+            ])
+        """
+
+    async def guard_text(
+        self,
+        text_or_messages: str | _T,
+        *,
+        recipe: str = "pangea_prompt_guard",
+        debug: bool = False,
+    ) -> PangeaResponse[TextGuardResult[_T]]:
+        """
+        Text Guard for scanning LLM inputs and outputs (Beta)
+
+        Analyze and redact text to avoid manipulation of the model, addition of
+        malicious content, and other undesirable data transfers.
+
+        How to install a [Beta release](https://pangea.cloud/docs/sdk/python/#beta-releases).
+
+        OperationId: ai_guard_post_v1beta_text_guard
+
+        Args:
+            text_or_messages: Text or structured data to be scanned by AI Guard
+                for PII, sensitive data, malicious content, and other data types
+                defined by the configuration. Supports processing up to 10KB of text.
+            recipe: Recipe key of a configuration of data types and settings
+                defined in the Pangea User Console. It specifies the rules that
+                are to be applied to the text, such as defang malicious URLs.
+            debug: Setting this value to true will provide a detailed analysis
+                of the text data
+        """
+
         return await self.request.post(
-            "v1beta/text/guard", TextGuardResult, data={"text": text, "recipe": recipe, "debug": debug}
+            "v1beta/text/guard",
+            TextGuardResult,
+            data={
+                "text" if isinstance(text_or_messages, str) else "messages": text_or_messages,
+                "recipe": recipe,
+                "debug": debug,
+            },
         )