+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Support for structured input to AI Guard.
+
 ## 5.4.0 - 2025-01-22
 
 ### Removed